dashboard/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-cloud-bigquery==3.17.2
gunicorn==22.0.0
